code/requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
arrow==1.3.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.3
boto3==1.37.27
botocore==1.37.27
bravado==11.1.0
bravado-core==6.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
cleantext==1.1.4
click==8.1.8
cut-cross-entropy==25.1.1
datasets==3.4.1
decorator==5.2.1
diffusers==0.32.2
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
hf_transfer==0.1.9
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipdb==0.13.13
ipython==8.34.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
neptune==1.13.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
s3transfer==0.11.4
safetensors==0.4.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
shtab==1.7.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
swagger-spec-validator==3.0.4
sympy==1.13.1
thefuzz==0.22.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
trl==0.15.2
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.1
unsloth==2025.3.15
unsloth_zoo==2025.3.13
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
websocket-client==1.8.0
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
